Add GENERATOR-DECOMPOSITION
This required major changes and surgery to the core group code. (The
core changes were actually recorded on YouTube, in a live-coding
exercise.) Essentially, as a perm group gets constructed, we keep track
of SLPs of the perms being passed around. We consequently store them
with the group, which allows decomposition of elements based off of
their generators.

There is a catch: When computing the generator decomposition, the length
of the produced sequence increases exponentially. (I think it is
exponential in the degree of the group, at least.) As such, it is not
very efficient.

In addition to the surgery, there was a lot of clean-up done. This isn't
good practice, but resulted from strenuous debugging of the code. I hope
that the code is actually easier to understand now in some places. The
clean-up includes extra documentation in the source files, too.